Refactor prompt manager and improve prompt handling
Removed `PromptManager` and related functionality, consolidating its logic into reusable utility functions in `util.go`. Updated tests to align with the refactored structure and introduced the `PromptGetResponse` type for cleaner response handling. Simplified prompt initialization and validation procedures.
